Adding ingredients
I created a new method that will be the handler when the user clicks on the more button. This method gets the old ingredient count by
getting the corresponding value for the type passed in the method, then it updates the count and updates the object. Finally,
it gets the ingredient price by the INGREDIENT_PRICES constant and add it to the totalPrice state.