Adding a Modal
The modal component is inside the UI components. This is used to display the order summary when the user clicks on the
order button. BurgerBuilder sets a new state (purchasing) as false, then, it uses a handler in order and connect it to
the order button. The Modal receives the purchasing state value in order to set some css inline style. If the purchasing
state is false, the modal is put at the top of the page and it has 0 of opacity, so that it is not visible. Then, when the user
clicks on the button, these css rules are now: translate the modal at position 0 and set opacity of one.