Adding backdrop
The backdrop component receives two props: isPurchasing in order to display the backdrop and the click handler, so that the user
can click on the backdrop, change the isPurchasing state. That handler is created as handlerModalClicked inside BurgerBuilder and
it passes that handler to the modal, so that it displays the backdrop and passes the state and the clicked handler.