Adding a custom UI Button component
This component renders a button and set the onClick handler as props.clicked (passed through the clicked prop when using the component),
then, it renders props.children inside the button and passes the variant in the className prop. I used this button in the OrderSummary
JS component receiving the method handler for each button as props.