Fetching orders
Now, I want to get the orders made by users from the firebase storage, so
I used axios and set the state of the Orders component as the response.data object.

Then, I pass the list of orders and the order information to the Order component that
unpacks the data and show it on some Material UI components.